backend/pyproject.toml
[project]
name = "ecommerce-backend"
version = "0.1.0"
description = "FastAPI backend for the ecommerce MVP"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",
    "supabase>=2.11.0",
    "pyjwt[crypto]>=2.10.0",
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.flake8-bugbear]
# FastAPI parameter markers are intentionally called in defaults.
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.Query",
    "fastapi.Path",
    "fastapi.Body",
    "fastapi.Header",
    "fastapi.Cookie",
    "fastapi.Form",
    "fastapi.File",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

# supabase-py returns typed as the full JSON union; we know the shapes at the
# call sites. Don't fight the SDK's generic types inside service modules.
[[tool.mypy.overrides]]
module = ["app.services.*"]
disable_error_code = ["arg-type", "return-value", "dict-item"]
